Create Sudoko_Solver.cpp
Quoting Wikipedia : "Sudoku is a logic-based, combinatorial number-placement puzzle. The objective is to fill a 9×9 grid so that each column, each row, and each of the nine 3×3 boxes (also called blocks or regions) contains the digits from 1 to 9 only one time each. The puzzle setter provides a partially completed grid." The rules for an N^2 X N^2 sudoku are as follows :

The board is consists of N^2 rows and N^2 columns.
Numbers between 1 and N^2(inclusive) are to be filled in each row such that :
All numbers in each row are distinct.
All numbers in each column are distinct.
All numbers in the sub-matrix having rows from (i*N + 1) to(i + 1)*N, and columns from (j*N + 1) to (j + 1)*N both inclusive, should be distinct. 0 <= i,j <= N-1. Rows and columns are 1 indexed. Each such sub-matrix is called a "box" or "region".
For this problem, you are required to solve a general N^2 X N^2 sudoku puzzle. Given a partially filled sudoku board, you have to fill it in as "perfect" a manner as possible.

INPUT
The first line should contains N. The following N lines contain 3 numbers: x, y and d. 1 <= x,y,d <= N^2. This means that a number d is present on the board at position (x,y)  0 <= N<= N^4 .All positions (x,y) in the input should be unique.

OUTPUT
Solved Sudoku if possible otherwise message indicating not solvable